Close #507: [`refined4s-doobie-ce2`] Reorganize type-class instances in the modules
