+++ modules/refined4s-doobie-ce2/shared/src/main/scala/refined4s/modules/doobie/derivation/types/strings.scala
@@ -1,0 +1,33 @@
+package refined4s.modules.doobie.derivation.types
+
+import cats.Show
+import doobie.{Get, Put}
+import refined4s.types.strings.*
+
+/** @author Kevin Lee
+  * @since 2025-09-02
+  */
+trait strings {
+
+  inline given derivedNonEmptyStringGet: Get[NonEmptyString] = strings.derivedNonEmptyStringGet
+  inline given derivedNonEmptyStringPut: Put[NonEmptyString] = strings.derivedNonEmptyStringPut
+
+  inline given derivedNonBlankStringGet: Get[NonBlankString] = strings.derivedNonBlankStringGet
+  inline given derivedNonBlankStringPut: Put[NonBlankString] = strings.derivedNonBlankStringPut
+
+  inline given derivedUuidGet: Get[Uuid] = strings.derivedUuidGet
+  inline given derivedUuidPut: Put[Uuid] = strings.derivedUuidPut
+
+}
+object strings {
+
+  given derivedNonEmptyStringGet(using Show[String]): Get[NonEmptyString] = Get[String].temap(NonEmptyString.from)
+  given derivedNonEmptyStringPut: Put[NonEmptyString]                     = Put[String].contramap(_.value)
+
+  given derivedNonBlankStringGet(using Show[String]): Get[NonBlankString] = Get[String].temap(NonBlankString.from)
+  given derivedNonBlankStringPut: Put[NonBlankString]                     = Put[String].contramap(_.value)
+
+  given derivedUuidGet(using Show[String]): Get[Uuid] = Get[String].temap(Uuid.from)
+  given derivedUuidPut: Put[Uuid]                     = Put[String].contramap(_.value)
+
+}
